Add new schedules to the DOM
We can now add new schedules by filling the form and submiting it,
validations and server side persistence is missing right now and should
be implemented next, as well as a mechanism to clear the form after the
submission. Preferentially without changing the dom directly but
instead, altering owner's state.