Add TestDatabase component in service-test
Added but the tests aren't running at all and I can't find out why. The
tests begin to fail if I attach the `test-system-interceptor` into the
`service-map` when creating a servlet.

Investigating on some issues on pedestal gh page I cought an idea to
create a server instead of a servlet for this tests.