Restarting the project!
After a little while, I'm going to put this project back in the rails.

A lot of stuff was done, dependencies updated, views changed, obsolete code was
commented out, etc.

I'll replacing Om with reagent and re-frame so two components needs to be
rewritten. The admin will be a full cljs app, with routes and all, it's not
going to be various html pages with a react component mounted.

Well, there is a lot to do, so let's get to work.