Schedule component is almost complete
Editing, deleting and creating is working, there is some minor adjustments to be
made though.

The component should be redesigned, I already have something in mind.

When editing an event the change is not getting reflected right away, you have
to change tabs to see the effects of it, when you delete an event from a tab,
all events just goes away and the page needs to be refreshed.

Then naming is a little confused, the `schedules` vars should be renamed to
`events`, so that a schedule is a collection of `events`. Maybe I'll have to
update the database as well.

All the subscriptions and and events must be namespaced, most of them already
are, but a few is missing namespaces.

The component's behavior could improve, create/edit an item on enter, also,
validations are not been shown to the user. Maybe a little bit of animation.