A little bit of refactoring and authentication.
Updated the yesql library so I had to change some namespaces that used the old
version of yesql api.

Created two tables: users and user_tokens as well as the auth namespace that
takes care of email and password authentication, token generation, storage and
provide an interceptor to use around our http service.

On the client side, there is a login page now which takes an email address and
password and send it to the server to get an access token, which will be used on
subsequent requests.

The token is saved to the local storage so we don't have to log in every time we
refresh the page. By default the token expires in 15 days and I'm not sure that
this approach is good enough, going to make some more research on this.

Active menu items are being colored according to the route now.

Refactor the routes namespace to use a re-frame/dispatch to load the schedule
instead of making the request and handle the results right there.